back-end/requirements.txt
fastapi
gunicorn
pyodbc
python-dotenv
SQLAlchemy
passlib
PyJWT
python-dateutil